+++ CHANGELOG.md
@@ -10,12 +10,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `batch_submit` task - [#41](https://github.com/PrefectHQ/prefect-aws/issues/41)
+- `MinIOCredentials` block - [#46](https://github.com/PrefectHQ/prefect-aws/pull/46)
 
 ### Changed
+
+- Converted `AwsCredentials` into a `Block` [#45](https://github.com/PrefectHQ/prefect-aws/pull/45)
 
 ### Deprecated
 
 ### Removed
+
+- Removed `.result()` and `is_complete` on test flow calls. [#45](https://github.com/PrefectHQ/prefect-aws/pull/45)
 
 ### Fixed